pyproject.toml
[project]
name = "pymarsbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "python-telegram-bot>=22.5",
]
